check to do phys analysis in stim-specific;
git-svn-id: svn://132.239.158.177/projects/ratrix/trunk@2504 d291aab3-510b-4049-a6d1-fe65b8d52d70

+++ setup/setProtocolTEST.m
@@ -346,7 +346,7 @@ fF = fullField(contrast,frequencies,duration,repetitions,maxWidth,maxHeight,scal
 
 % manualCmrMotionEyeCal
 background=0.5;
-numSweeps=5;
+numSweeps=uint8(5);
 maxWidth=1024;
 maxHeight=768;
 manualCam = manualCmrMotionEyeCal(background,numSweeps,maxWidth,maxHeight,scaleFactor,interTrialLuminance);
